requirements.txt
astroid==2.4.2
attrs==19.3.0
blessings==1.7
bpython==0.18
captcha==0.3
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.1
Django==2.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-ranged-response==0.2.0
django-shell-plus==1.1.7
django-simple-captcha==0.5.12
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
Pygments==2.5.2
pylint==2.5.3
pyparsing==2.4.6
pytest==5.4.3
python-dotenv==0.13.0
pytz==2019.3
requests==2.22.0
sentry-sdk==0.16.2
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
toml==0.10.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.12.1
zipp==3.0.0
